+++ tapgarden/mock.go
@@ -1,39 +1,27 @@
 package tapgarden
 
 import (
-	"bytes"
 	"context"
-	"crypto/sha256"
 	"encoding/hex"
 	"fmt"
-	"math/rand"
-	"sync"
 	"sync/atomic"
 	"testing"
 	"time"
 
-	btcaddr "github.com/btcsuite/btcd/address/v2"
-	"github.com/btcsuite/btcd/btcec/v2"
 	"github.com/btcsuite/btcd/btcec/v2/schnorr"
-	"github.com/btcsuite/btcd/chaincfg/v2"
 	"github.com/btcsuite/btcd/chainhash/v2"
 	"github.com/btcsuite/btcd/psbt/v2"
-	"github.com/btcsuite/btcd/txscript/v2"
 	"github.com/btcsuite/btcd/wire/v2"
-	"github.com/btcsuite/btcwallet/waddrmgr"
 	"github.com/lightninglabs/lndclient"
 	"github.com/lightninglabs/taproot-assets/address"
 	"github.com/lightninglabs/taproot-assets/asset"
 	"github.com/lightninglabs/taproot-assets/fn"
 	"github.com/lightninglabs/taproot-assets/internal/test"
 	"github.com/lightninglabs/taproot-assets/proof"
-	"github.com/lightninglabs/taproot-assets/tapscript"
+	"github.com/lightninglabs/taproot-assets/tapnode/tapnodemock"
 	"github.com/lightninglabs/taproot-assets/tapsend"
-	"github.com/lightningnetwork/lnd/chainntnfs"
 	"github.com/lightningnetwork/lnd/keychain"
-	"github.com/lightningnetwork/lnd/lnwallet"
 	"github.com/lightningnetwork/lnd/lnwallet/chainfee"
-	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 )
 
@@ -278,7 +266,7 @@ func RandMintingBatch(t testing.TB, opts ...MintBatchOption) *MintingBatch {
 	walletFundPsbt := func(ctx context.Context,
 		anchorPkt psbt.Packet) (tapsend.FundedPsbt, error) {
 
-		changeOutputIdx := FundGenesisTx(
+		changeOutputIdx := tapnodemock.FundGenesisTx(
 			&anchorPkt, chainfee.FeePerKwFloor,
 		)
 
@@ -321,518 +309,6 @@ func RandSeedlings(t testing.TB, numSeedlings int) map[string]*Seedling {
 	}
 
 	return seedlings
-}
-
-type MockWalletAnchor struct {
-	FundPsbtSignal     chan *tapsend.FundedPsbt
-	SignPsbtSignal     chan struct{}
-	ImportPubKeySignal chan *btcec.PublicKey
-	ListUnspentSignal  chan struct{}
-	SubscribeTxSignal  chan struct{}
-	SubscribeTx        chan lndclient.Transaction
-	ListTxnsSignal     chan struct{}
-
-	Transactions  []lndclient.Transaction
-	ImportedUtxos []*lnwallet.Utxo
-}
-
-func NewMockWalletAnchor() *MockWalletAnchor {
-	return &MockWalletAnchor{
-		FundPsbtSignal:     make(chan *tapsend.FundedPsbt),
-		SignPsbtSignal:     make(chan struct{}),
-		ImportPubKeySignal: make(chan *btcec.PublicKey),
-		ListUnspentSignal:  make(chan struct{}),
-		SubscribeTxSignal:  make(chan struct{}),
-		SubscribeTx:        make(chan lndclient.Transaction),
-		ListTxnsSignal:     make(chan struct{}),
-	}
-}
-
-// NewGenesisTx creates a funded genesis PSBT with the given fee rate.
-func NewGenesisTx(t testing.TB, feeRate chainfee.SatPerKWeight) psbt.Packet {
-	txTemplate := wire.NewMsgTx(2)
-	txTemplate.AddTxOut(tapsend.CreateDummyOutput())
-	genesisPkt, err := psbt.NewFromUnsignedTx(txTemplate)
-	require.NoError(t, err)
-
-	FundGenesisTx(genesisPkt, feeRate)
-	return *genesisPkt
-}
-
-// FundGenesisTx add a genesis input and change output to a 1-output TX and
-// returns the index of the change output.
-func FundGenesisTx(packet *psbt.Packet, feeRate chainfee.SatPerKWeight) uint32 {
-	const anchorBalance = int64(100000)
-
-	// Take the PSBT packet and add an additional input and output to
-	// simulate the wallet funding the transaction.
-	packet.UnsignedTx.AddTxIn(&wire.TxIn{
-		PreviousOutPoint: wire.OutPoint{
-			Index: test.RandInt[uint32](),
-		},
-	})
-
-	// Use a P2TR input by default.
-	anchorInput := psbt.PInput{
-		WitnessUtxo: &wire.TxOut{
-			Value:    anchorBalance,
-			PkScript: bytes.Clone(tapsend.GenesisDummyScript),
-		},
-		SighashType: txscript.SigHashDefault,
-	}
-	packet.Inputs = append(packet.Inputs, anchorInput)
-
-	// Use a non-P2TR change output by default so we avoid generating
-	// exclusion proofs.
-	changeOutput := wire.TxOut{
-		Value:    anchorBalance - packet.UnsignedTx.TxOut[0].Value,
-		PkScript: bytes.Clone(tapsend.GenesisDummyScript),
-	}
-	changeOutput.PkScript[0] = txscript.OP_0
-	packet.UnsignedTx.AddTxOut(&changeOutput)
-	packet.Outputs = append(packet.Outputs, psbt.POutput{})
-
-	// Set a realistic change value.
-	_, fee := tapscript.EstimateFee(
-		[][]byte{tapsend.GenesisDummyScript}, packet.UnsignedTx.TxOut,
-		feeRate,
-	)
-	changeOutputIdx := len(packet.UnsignedTx.TxOut) - 1
-	packet.UnsignedTx.TxOut[changeOutputIdx].Value -= int64(fee)
-
-	return uint32(changeOutputIdx)
-}
-
-// FundPsbt funds a PSBT.
-func (m *MockWalletAnchor) FundPsbt(_ context.Context, packet *psbt.Packet,
-	_ uint32, _ chainfee.SatPerKWeight,
-	changeIdx int32) (*tapsend.FundedPsbt, error) {
-
-	// Take the PSBT packet and add an additional input and output to
-	// simulate the wallet funding the transaction.
-	packet.UnsignedTx.AddTxIn(&wire.TxIn{
-		PreviousOutPoint: wire.OutPoint{
-			Index: rand.Uint32(),
-		},
-	})
-
-	// Use a P2TR input by default.
-	anchorInput := psbt.PInput{
-		WitnessUtxo: &wire.TxOut{
-			Value:    100000,
-			PkScript: bytes.Clone(tapsend.GenesisDummyScript),
-		},
-		SighashType: txscript.SigHashDefault,
-	}
-	packet.Inputs = append(packet.Inputs, anchorInput)
-
-	// Use a non-P2TR change output by default so we avoid generating
-	// exclusion proofs.
-	changeOutput := wire.TxOut{
-		Value:    50000,
-		PkScript: bytes.Clone(tapsend.GenesisDummyScript),
-	}
-	changeOutput.PkScript[0] = txscript.OP_0
-	packet.UnsignedTx.AddTxOut(&changeOutput)
-	packet.Outputs = append(packet.Outputs, psbt.POutput{})
-
-	// The change output was added last, so it will be the last output in
-	// the list. Update the change index to reflect this.
-	changeIdx = int32(len(packet.Outputs) - 1)
-
-	// We always have the change output be the second output, so this means
-	// the Taproot Asset commitment will live in the first output.
-	pkt := &tapsend.FundedPsbt{
-		Pkt:               packet,
-		ChangeOutputIndex: changeIdx,
-	}
-
-	m.FundPsbtSignal <- pkt
-
-	return pkt, nil
-}
-
-func (m *MockWalletAnchor) SignAndFinalizePsbt(ctx context.Context,
-	pkt *psbt.Packet) (*psbt.Packet, error) {
-
-	select {
-	case <-ctx.Done():
-		return nil, fmt.Errorf("shutting down")
-	default:
-	}
-
-	// We'll modify the packet by attaching a "signature" so the PSBT
-	// appears to actually be finalized.
-	pkt.Inputs[0].FinalScriptSig = []byte{}
-
-	select {
-	case <-ctx.Done():
-		return nil, fmt.Errorf("shutting down")
-	case m.SignPsbtSignal <- struct{}{}:
-	}
-
-	return pkt, nil
-}
-
-func (m *MockWalletAnchor) ImportTaprootOutput(ctx context.Context,
-	pub *btcec.PublicKey) (btcaddr.Address, error) {
-
-	select {
-	case m.ImportPubKeySignal <- pub:
-
-	case <-ctx.Done():
-		return nil, fmt.Errorf("shutting down")
-	}
-
-	return btcaddr.NewAddressTaproot(
-		schnorr.SerializePubKey(pub), &chaincfg.RegressionNetParams,
-	)
-}
-
-// UnlockInput unlocks the set of target inputs after a batch or send
-// transaction is abandoned.
-func (m *MockWalletAnchor) UnlockInput(context.Context, wire.OutPoint) error {
-	return nil
-}
-
-// ListUnspentImportScripts lists all UTXOs of the imported Taproot scripts.
-func (m *MockWalletAnchor) ListUnspentImportScripts(
-	ctx context.Context) ([]*lnwallet.Utxo, error) {
-
-	select {
-	case m.ListUnspentSignal <- struct{}{}:
-
-	case <-ctx.Done():
-		return nil, fmt.Errorf("shutting down")
-	}
-
-	return m.ImportedUtxos, nil
-}
-
-// ImportTapscript imports a Taproot output script into the wallet to track it
-// on-chain in a watch-only manner.
-func (m *MockWalletAnchor) ImportTapscript(_ context.Context,
-	tapscript *waddrmgr.Tapscript) (btcaddr.Address, error) {
-
-	taprootKey, err := tapscript.TaprootKey()
-	if err != nil {
-		return nil, err
-	}
-
-	return btcaddr.NewAddressTaproot(
-		schnorr.SerializePubKey(taprootKey),
-		&chaincfg.RegressionNetParams,
-	)
-}
-
-// SubscribeTransactions creates a uni-directional stream from the server to the
-// client in which any newly discovered transactions relevant to the wallet are
-// sent over.
-func (m *MockWalletAnchor) SubscribeTransactions(
-	ctx context.Context) (<-chan lndclient.Transaction, <-chan error, error) {
-
-	select {
-	case m.SubscribeTxSignal <- struct{}{}:
-
-	case <-ctx.Done():
-		return nil, nil, fmt.Errorf("shutting down")
-	}
-
-	errChan := make(chan error)
-	return m.SubscribeTx, errChan, nil
-}
-
-// ListTransactions returns all known transactions of the backing lnd node. It
-// takes a start and end block height which can be used to limit the block range
-// that we query over. These values can be left as zero to include all blocks.
-// To include unconfirmed transactions in the query, endHeight must be set to
-// -1.
-func (m *MockWalletAnchor) ListTransactions(ctx context.Context, _, _ int32,
-	_ string) ([]lndclient.Transaction, error) {
-
-	select {
-	case m.ListTxnsSignal <- struct{}{}:
-
-	case <-ctx.Done():
-		return nil, fmt.Errorf("shutting down")
-	}
-
-	return m.Transactions, nil
-}
-
-// MinRelayFee estimates the minimum fee rate required for a
-// transaction.
-func (m *MockWalletAnchor) MinRelayFee(
-	ctx context.Context) (chainfee.SatPerKWeight, error) {
-
-	return chainfee.SatPerKWeight(10), nil
-}
-
-type MockChainBridge struct {
-	FeeEstimateSignal chan struct{}
-	PublishReq        chan *wire.MsgTx
-	ConfReqSignal     chan int
-	BlockEpochSignal  chan struct{}
-
-	NewBlocks chan int32
-
-	ReqCount atomic.Int32
-
-	// ConfReqs and confErr are not guarded by a mutex. Readers must
-	// synchronise with the writer via the ConfReqSignal channel: the
-	// writer (RegisterConfirmationsNtfn) populates ConfReqs and
-	// reassigns confErr before sending the request number on
-	// ConfReqSignal, so any receive from that channel happens-after
-	// the corresponding store and a read of ConfReqs[reqNo] is safe.
-	ConfReqs map[int]*chainntnfs.ConfirmationEvent
-
-	// BlocksMu protects concurrent access to Blocks. Readers (GetBlock,
-	// called from caretaker goroutines) hold it for read; all writers
-	// must go through SetBlock so the invariant cannot be violated
-	// piecemeal.
-	BlocksMu sync.RWMutex
-	Blocks   map[chainhash.Hash]*wire.MsgBlock
-
-	failFeeEstimates atomic.Bool
-	errConf          atomic.Int32
-	emptyConf        atomic.Int32
-	confErr          chan error
-}
-
-func NewMockChainBridge() *MockChainBridge {
-	return &MockChainBridge{
-		FeeEstimateSignal: make(chan struct{}),
-		PublishReq:        make(chan *wire.MsgTx),
-		ConfReqs:          make(map[int]*chainntnfs.ConfirmationEvent),
-		ConfReqSignal:     make(chan int),
-		BlockEpochSignal:  make(chan struct{}, 1),
-		NewBlocks:         make(chan int32),
-		Blocks:            make(map[chainhash.Hash]*wire.MsgBlock),
-	}
-}
-
-func (m *MockChainBridge) FailFeeEstimatesOnce() {
-	m.failFeeEstimates.Store(true)
-}
-
-// FailConfOnce updates the ChainBridge such that the next call to
-// RegisterConfirmationNtfn will fail by returning an error on the error channel
-// returned from RegisterConfirmationNtfn.
-func (m *MockChainBridge) FailConfOnce() {
-	// Store the incremented request count so we never store 0 as a value.
-	m.errConf.Store(m.ReqCount.Load() + 1)
-}
-
-// EmptyConfOnce updates the ChainBridge such that the next confirmation event
-// sent via SendConfNtfn will have an empty confirmation.
-func (m *MockChainBridge) EmptyConfOnce() {
-	// Store the incremented request count so we never store 0 as a value.
-	m.emptyConf.Store(m.ReqCount.Load() + 1)
-}
-
-func (m *MockChainBridge) SendConfNtfn(reqNo int, blockHash *chainhash.Hash,
-	blockHeight, blockIndex int, block *wire.MsgBlock,
-	tx *wire.MsgTx) {
-
-	// Compare to the incremented request count since we incremented it
-	// when storing the request number.
-	req := m.ConfReqs[reqNo]
-	if m.emptyConf.Load() == int32(reqNo)+1 {
-		m.emptyConf.Store(0)
-		req.Confirmed <- nil
-		return
-	}
-
-	req.Confirmed <- &chainntnfs.TxConfirmation{
-		BlockHash:   blockHash,
-		BlockHeight: uint32(blockHeight),
-		TxIndex:     uint32(blockIndex),
-		Block:       block,
-		Tx:          tx,
-	}
-}
-
-func (m *MockChainBridge) RegisterConfirmationsNtfn(ctx context.Context,
-	_ *chainhash.Hash, _ []byte, _, _ uint32, _ bool,
-	_ chan struct{}) (*chainntnfs.ConfirmationEvent, chan error, error) {
-
-	select {
-	case <-ctx.Done():
-		return nil, nil, fmt.Errorf("shutting down")
-	default:
-	}
-
-	defer func() {
-		m.ReqCount.Add(1)
-	}()
-
-	req := &chainntnfs.ConfirmationEvent{
-		Confirmed: make(chan *chainntnfs.TxConfirmation),
-		Cancel:    func() {},
-	}
-	m.confErr = make(chan error, 1)
-
-	currentReqCount := m.ReqCount.Load()
-	m.ConfReqs[int(currentReqCount)] = req
-
-	select {
-	case m.ConfReqSignal <- int(currentReqCount):
-	case <-ctx.Done():
-	}
-
-	// Compare to the incremented request count since we incremented it
-	// when storing the request number.
-	if m.errConf.CompareAndSwap(currentReqCount+1, 0) {
-		m.confErr <- fmt.Errorf("confirmation registration error")
-	}
-
-	return req, m.confErr, nil
-}
-
-func (m *MockChainBridge) RegisterBlockEpochNtfn(
-	ctx context.Context) (chan int32, chan error, error) {
-
-	select {
-	case <-ctx.Done():
-		return nil, nil, fmt.Errorf("shutting down")
-	default:
-	}
-
-	select {
-	case m.BlockEpochSignal <- struct{}{}:
-	case <-ctx.Done():
-	}
-
-	return m.NewBlocks, make(chan error), nil
-}
-
-// GetBlock returns a chain block given its hash.
-func (m *MockChainBridge) GetBlock(ctx context.Context,
-	hash chainhash.Hash) (*wire.MsgBlock, error) {
-
-	m.BlocksMu.RLock()
-	block, ok := m.Blocks[hash]
-	m.BlocksMu.RUnlock()
-	if !ok {
-		return nil, fmt.Errorf("block %s not found", hash.String())
-	}
-
-	return block, nil
-}
-
-// SetBlock records a block under its hash so a later GetBlock can return
-// it. All writers to Blocks must go through this helper so the BlocksMu
-// invariant cannot be violated piecemeal.
-func (m *MockChainBridge) SetBlock(hash chainhash.Hash, block *wire.MsgBlock) {
-	m.BlocksMu.Lock()
-	m.Blocks[hash] = block
-	m.BlocksMu.Unlock()
-}
-
-// GetBlockByHeight returns a block given the block height.
-func (m *MockChainBridge) GetBlockByHeight(ctx context.Context,
-	blockHeight int64) (*wire.MsgBlock, error) {
-
-	return &wire.MsgBlock{}, nil
-}
-
-// GetBlockHeaderByHeight returns a block header given the block height.
-func (m *MockChainBridge) GetBlockHeaderByHeight(ctx context.Context,
-	blockHeight int64) (*wire.BlockHeader, error) {
-
-	return &wire.BlockHeader{}, nil
-}
-
-// GetBlockHash returns the hash of the block in the best blockchain at the
-// given height.
-func (m *MockChainBridge) GetBlockHash(ctx context.Context,
-	blockHeight int64) (chainhash.Hash, error) {
-
-	return chainhash.Hash{}, nil
-}
-
-// VerifyBlock returns an error if a block (with given header and height) is not
-// present on-chain. It also checks to ensure that block height corresponds to
-// the given block header.
-func (m *MockChainBridge) VerifyBlock(_ context.Context,
-	_ wire.BlockHeader, _ uint32) error {
-
-	return nil
-}
-
-func (m *MockChainBridge) CurrentHeight(_ context.Context) (uint32, error) {
-	return 0, nil
-}
-
-func (m *MockChainBridge) GetBlockTimestamp(_ context.Context, _ uint32) (int64,
-	error) {
-
-	return 0, nil
-}
-
-func (m *MockChainBridge) PublishTransaction(_ context.Context,
-	tx *wire.MsgTx, _ string) error {
-
-	m.PublishReq <- tx
-	return nil
-}
-
-func (m *MockChainBridge) EstimateFee(ctx context.Context,
-	_ uint32) (chainfee.SatPerKWeight, error) {
-
-	select {
-	case m.FeeEstimateSignal <- struct{}{}:
-
-	case <-ctx.Done():
-		return 0, fmt.Errorf("shutting down")
-	}
-
-	if m.failFeeEstimates.Load() {
-		m.failFeeEstimates.Store(false)
-		return 0, fmt.Errorf("failed to estimate fee")
-	}
-
-	return chainfee.FeePerKwFloor, nil
-}
-
-// TxBlockHeight returns the block height that the given transaction was
-// included in.
-func (m *MockChainBridge) TxBlockHeight(context.Context,
-	chainhash.Hash) (uint32, error) {
-
-	return 123, nil
-}
-
-// MeanBlockTimestamp returns the timestamp of the block at the given height as
-// a Unix timestamp in seconds, taking into account the mean time elapsed over
-// the previous 11 blocks.
-func (m *MockChainBridge) MeanBlockTimestamp(context.Context,
-	uint32) (time.Time, error) {
-
-	return time.Now(), nil
-}
-
-// GenFileChainLookup generates a chain lookup interface for the given
-// proof file that can be used to validate proofs.
-func (m *MockChainBridge) GenFileChainLookup(*proof.File) asset.ChainLookup {
-	return m
-}
-
-// GenProofChainLookup generates a chain lookup interface for the given
-// single proof that can be used to validate proofs.
-func (m *MockChainBridge) GenProofChainLookup(*proof.Proof) (asset.ChainLookup,
-	error) {
-
-	return m, nil
-}
-
-var _ asset.ChainLookup = (*MockChainBridge)(nil)
-var _ ChainBridge = (*MockChainBridge)(nil)
-
-func GenMockGroupVerifier() func(*btcec.PublicKey) error {
-	return func(groupKey *btcec.PublicKey) error {
-		return nil
-	}
 }
 
 type MockAssetSyncer struct {
@@ -906,188 +382,12 @@ func (m *MockAssetSyncer) EnableAssetSync(_ context.Context,
 	return nil
 }
 
-type MockKeyRing struct {
-	mock.Mock
-
-	sync.RWMutex
-
-	KeyIndex uint32
-
-	Keys map[keychain.KeyLocator]*btcec.PrivateKey
-
-	// deriveNextKeyCallCount is used to track the number of calls to
-	// DeriveNextKey.
-	deriveNextKeyCallCount atomic.Uint64
-}
-
-var _ KeyRing = (*MockKeyRing)(nil)
-
-func NewMockKeyRing() *MockKeyRing {
-	keyRing := &MockKeyRing{
-		Keys: make(map[keychain.KeyLocator]*btcec.PrivateKey),
-	}
-
-	keyRing.On(
-		"DeriveNextKey", mock.Anything,
-		keychain.KeyFamily(asset.TaprootAssetsKeyFamily),
-	).Return(keychain.KeyDescriptor{}, nil)
-
-	keyRing.On(
-		"DeriveNextTaprootAssetKey", mock.Anything,
-	).Return(keychain.KeyDescriptor{}, nil)
-
-	return keyRing
-}
-
-// DeriveNextTaprootAssetKey attempts to derive the *next* key within the
-// Taproot Asset key family.
-func (m *MockKeyRing) DeriveNextTaprootAssetKey(
-	ctx context.Context) (keychain.KeyDescriptor, error) {
-
-	// No need to lock mutex here, DeriveNextKey does that for us.
-	m.Called(ctx)
-
-	return m.DeriveNextKey(ctx, asset.TaprootAssetsKeyFamily)
-}
-
-func (m *MockKeyRing) DeriveNextKey(ctx context.Context,
-	keyFam keychain.KeyFamily) (keychain.KeyDescriptor, error) {
-
-	m.Lock()
-	defer func() {
-		m.KeyIndex++
-		m.Unlock()
-	}()
-
-	m.Called(ctx, keyFam)
-	m.deriveNextKeyCallCount.Add(1)
-
-	select {
-	case <-ctx.Done():
-		return keychain.KeyDescriptor{}, fmt.Errorf("shutting down")
-	default:
-	}
-
-	priv, err := btcec.NewPrivateKey()
-	if err != nil {
-		return keychain.KeyDescriptor{}, err
-	}
-
-	loc := keychain.KeyLocator{
-		Index:  m.KeyIndex,
-		Family: keyFam,
-	}
-
-	m.Keys[loc] = priv
-
-	desc := keychain.KeyDescriptor{
-		PubKey:     priv.PubKey(),
-		KeyLocator: loc,
-	}
-
-	return desc, nil
-}
-
-func (m *MockKeyRing) IsLocalKey(ctx context.Context,
-	d keychain.KeyDescriptor) bool {
-
-	m.Lock()
-	defer m.Unlock()
-
-	m.Called(ctx, d)
-
-	priv, ok := m.Keys[d.KeyLocator]
-	if ok && priv.PubKey().IsEqual(d.PubKey) {
-		return true
-	}
-
-	for _, key := range m.Keys {
-		if key.PubKey().IsEqual(d.PubKey) {
-			return true
-		}
-	}
-
-	return false
-}
-
-func (m *MockKeyRing) PubKeyAt(t *testing.T, idx uint32) *btcec.PublicKey {
-	m.Lock()
-	defer m.Unlock()
-
-	loc := keychain.KeyLocator{
-		Index:  idx,
-		Family: asset.TaprootAssetsKeyFamily,
-	}
-
-	priv, ok := m.Keys[loc]
-	if !ok {
-		t.Fatalf("script key not found at index %d", idx)
-	}
-
-	return priv.PubKey()
-}
-
-func (m *MockKeyRing) ScriptKeyAt(t *testing.T, idx uint32) asset.ScriptKey {
-	m.Lock()
-	defer m.Unlock()
-
-	loc := keychain.KeyLocator{
-		Index:  idx,
-		Family: asset.TaprootAssetsKeyFamily,
-	}
-
-	priv, ok := m.Keys[loc]
-	if !ok {
-		t.Fatalf("script key not found at index %d", idx)
-	}
-
-	return asset.NewScriptKeyBip86(keychain.KeyDescriptor{
-		KeyLocator: loc,
-		PubKey:     priv.PubKey(),
-	})
-}
-
-func (m *MockKeyRing) DeriveSharedKey(_ context.Context, key *btcec.PublicKey,
-	locator *keychain.KeyLocator) ([sha256.Size]byte, error) {
-
-	m.Lock()
-	defer m.Unlock()
-
-	if locator == nil {
-		return [32]byte{}, fmt.Errorf("locator is nil")
-	}
-
-	priv, ok := m.Keys[*locator]
-	if !ok {
-		return [32]byte{}, fmt.Errorf("script key not found at index "+
-			"%d", locator.Index)
-	}
-
-	ecdh := &keychain.PrivKeyECDH{
-		PrivKey: priv,
-	}
-	return ecdh.ECDH(key)
-}
-
-// DeriveNextKeyCallCount returns the number of calls to DeriveNextKey. This is
-// useful in tests to assert that the key ring was used as expected in
-// concurrent scenarios.
-func (m *MockKeyRing) DeriveNextKeyCallCount() int {
-	return int(m.deriveNextKeyCallCount.Load())
-}
-
-// ResetDeriveNextKeyCallCount resets the call counter for DeriveNextKey to
-// zero. This is useful in tests to ensure a clean state for assertions.
-func (m *MockKeyRing) ResetDeriveNextKeyCallCount() {
-	m.deriveNextKeyCallCount.Store(0)
-}
-
 type MockGenSigner struct {
-	KeyRing     *MockKeyRing
+	KeyRing     *tapnodemock.KeyRing
 	failSigning atomic.Bool
 }
 
-func NewMockGenSigner(keyRing *MockKeyRing) *MockGenSigner {
+func NewMockGenSigner(keyRing *tapnodemock.KeyRing) *MockGenSigner {
 	return &MockGenSigner{
 		KeyRing: keyRing,
 	}
